refactor(sync): extract LAN discovery into peers/lan.rs
